first pass at refactoring, main command function signatures simpler
