+++ commands/createApp.js
@@ -1,16 +1,18 @@
 const inquirer = require('inquirer')
 const path = require('path')
-const fs = require('fs')
 const mkdirp = require('mkdirp')
 const parallel = require('run-parallel')
-const bulk = require('bulk-require')
-const map = require('lodash/map')
 const series = require('run-series')
-const exec = require('child_process').exec
+const bulk = require('bulk-require')
 
 const createTopic = require('./createTopic')
 const coreTemplatesDir = path.resolve(__dirname, '../templates/core')
 const coreTemplates = bulk(coreTemplatesDir, '**/*.js')
+
+const createCore = require('./helpers/createCore')
+const appendToAppTopic = require('./helpers/appendToAppTopic')
+const runMigrations = require('./helpers/runMigrations')
+const installPackages = require('./helpers/installPackages')
 
 /*
 - Creates a folder for a dogstack app
@@ -56,17 +58,19 @@ module.exports = function createApp ({ appName, appDir, doInstallPackages, doRun
         if (err) {
           cb(err)
         } else {
-          const topicName = 'dogs'
+          // create the core / top-level folders and files, then the app topic and a starter topic
+          // then append to the app topic with extras it needs, and install packages / run migrations if user asked for them
+          const starterTopicName = 'dogs'
           series([
             createCore({ appName, dir: appDir, templates: coreTemplates }),
             function (cb) {
               parallel([
-                createTopic({ topicName: 'app', appDir, dir: path.join(appDir, 'app') }),
-                createTopic({ topicName, appDir, dir: path.join(appDir, topicName) })
+                createTopic({ topicName: 'app', topicDir: path.join(appDir, 'app'), appDir }),
+                createTopic({ topicName: starterTopicName, topicDir: path.join(appDir, starterTopicName), appDir })
               ], cb)
             },
             function (cb) {
-              const extras = [appendToAppTopic({ dir: path.join(appDir, 'app') })]
+              const extras = [ appendToAppTopic({ dir: path.join(appDir, 'app') }) ]
               if (doInstallPackages) extras.push(installPackages({ appDir }))
               if (doRunMigrations) extras.push(runMigrations({ appDir, appName }))
               series(extras, cb)
@@ -74,119 +78,6 @@ module.exports = function createApp ({ appName, appDir, doInstallPackages, doRun
           ], cb)
         }
       })
-    }
-  }
-}
-
-function createCore ({ appName, dir, templates }) {
-  return function (cb) {
-    parallel(
-      map(templates, (templateFn, name) => {
-        return function (cb) {
-          // TODO: IK: a more elegant way of guarding this
-          if (Object.keys(templateFn).length > 0 && name !== 'index') {
-            // a folder with core files at some sub-level
-            const subDir = path.join(dir, name)
-            mkdirp(subDir, function (err) {
-              if (err) {
-                cb(err)
-              } else {
-                createCore({ appName, dir: subDir, templates: templateFn })(cb)
-              }
-            })
-          } else {
-            // TODO: IK: a more robust way of identifying files that aren't .js, some regex or something
-            // cause what about .test.js files in the future?
-            const filename = name.includes('.') ? name : `${name}.js`
-            const filePath = path.join(dir, filename)
-            const template = templateFn(appName)
-            fs.writeFile(filePath, template, function (err) {
-              if (err) return cb(err)
-              cb(null, name)
-            })
-          }
-        }
-      }),
-      cb
-    )
-  }
-}
-
-function installPackages ({ appDir }) {
-  return function (cb) {
-    console.log('Installing packages...')
-    try {
-      const prevDir = process.cwd()
-      process.chdir(appDir)
-      exec('npm install', function (err) {
-        if (err) {
-          cb(err)
-        } else {
-          process.chdir(prevDir)
-          cb(null, 'Packages installed successfully!')
-        }
-      })
-    } catch (err) {
-      cb(err)
-    }
-  }
-}
-
-function runMigrations ({ appDir, appName }) {
-  return function (cb) {
-    console.log('Running migrations...')
-    const prevDir = process.cwd()
-    process.chdir(appDir)
-    exec(`psql -c "CREATE DATABASE ${appName}_development"`, function (err, stdout, stderr) {
-      // TODO: IK: a better way of handling this
-      const dbExistsError = `ERROR:  database "${appName}_development" already exists\n`
-      if (err && stderr !== dbExistsError) {
-        cb(err)
-      } else {
-        exec(`npm run db migrate:latest`, function (err) {
-          if (err) {
-            cb(err)
-          } else {
-            process.chdir(prevDir)
-            cb(null, 'Migrations run successfully!')
-          }
-        })
-      }
-    })
-  }
-}
-
-// TODO: IK: this is broken out to allow creating the 'app' topic to use createTopic
-// consider whether these folders / files might be better somewhere else in future
-function appendToAppTopic ({ dir }) {
-  const appTemplatesDir = path.resolve(__dirname, '../templates/app')
-  const appTemplates = bulk(appTemplatesDir, '**/*.js')
-
-  return function (cb) {
-    parallel([
-      createCore({ appName: null, dir, templates: appTemplates }),
-      copyNonJS({ dir })
-    ], cb)
-  }
-
-  function copyNonJS ({ dir }) {
-    return function (cb) {
-      // TODO: IK: only handling the favicon hardcoded atm, maybe theres also a better way to do this
-      const file = 'favicon.ico'
-      const reader = fs.createReadStream(path.join(appTemplatesDir, file))
-      const writer = fs.createWriteStream(path.join(dir, file))
-      writer.on('finish', function () {
-        cb(null)
-      })
-
-      reader.on('error', function (err) {
-        cb(err)
-      })
-      writer.on('error', function (err) {
-        cb(err)
-      })
-
-      reader.pipe(writer)
     }
   }
 }
