+++ templates/types/migration.js
@@ -1,3 +1,5 @@
+// TODO: IK: i'd rather not have this in types, but maybe clean up when also cleaning up 'core' and 'app' templates folders
+
 const dedent = require('dedent')
 
 module.exports = function (topic) {
